+++ main.ts
@@ -1,9 +1,10 @@
 import "./paths";
 import { config } from "@packages/config";
 import { app } from "@packages/backend";
+import { logger } from "@packages/logger";
 
 app.listen(config.server.port, () => {
-  console.log(
-    `🚀 http://${config.server.host}:${config.server.port} (${config.env})`,
+  logger.info(
+    `🚀 http://${config.server.host}:${config.server.port} (${config.mode})`,
   );
 });