+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Monorepo paths resolution
 - Type-safe configuration with dotenv / env-var
 - Fastify backend
+- Winston logger